Fixed Victory-Native dependency to version 13.4.0
Expo has precompiled code for some svg components. This conflicts with the dependencies of Victory-Native and in recent versions the module completely crashes on Android. For now, the only fix is to prevent updates of the package and stick with version 13.4.0. Another solution could be to manually modify the precompiled files of the expo managed system but it would probably require a lot of work and would be difficult to maintain.